refactor: split up code
