+++ src/restapi/analysis.rs
@@ -1,0 +1,13 @@
+use goose::goose::{GooseUser, TransactionResult};
+
+pub async fn get_analysis_status(user: &mut GooseUser) -> TransactionResult {
+    let _response = user.get("/api/v2/analysis/status").await?;
+
+    Ok(())
+}
+
+pub async fn get_analysis_latest_cpe(user: &mut GooseUser) -> TransactionResult {
+    let _response = user.get("/api/v2/analysis/latest/component/cpe%3A%2Fa%3Aredhat%3Aopenshift_builds%3A1.3%3A%3Ael9").await?;
+
+    Ok(())
+}
