+++ src/restapi/mod.rs
@@ -1,0 +1,6 @@
+pub mod advisory;
+pub mod analysis;
+pub mod misc;
+pub mod purl;
+pub mod sbom;
+pub mod vulnerability;
