+++ src/restapi/vulnerability.rs
@@ -1,0 +1,37 @@
+use goose::goose::{GooseUser, TransactionResult};
+use serde_json::json;
+
+pub async fn list_vulnerabilities(user: &mut GooseUser) -> TransactionResult {
+    let _response = user.get("/api/v2/vulnerability").await?;
+
+    Ok(())
+}
+
+pub async fn list_vulnerabilities_paginated(user: &mut GooseUser) -> TransactionResult {
+    let _response = user
+        .get("/api/v2/vulnerability?offset=100&limit=10")
+        .await?;
+
+    Ok(())
+}
+
+pub async fn get_vulnerability(id: String, user: &mut GooseUser) -> TransactionResult {
+    let _response = user.get(&format!("/api/v2/vulnerability/{id}")).await?;
+
+    Ok(())
+}
+
+pub async fn post_vulnerability_analyze(purl: String, user: &mut GooseUser) -> TransactionResult {
+    let _response = user
+        .post_json(
+            "/api/v2/vulnerability/analyze",
+            &json!({
+                "purls": [
+                     purl
+                ]
+            }),
+        )
+        .await?;
+
+    Ok(())
+}
